Move expression parsing code under new query system
The query expression parser is now only used by the new query system, so move it adjacent to the rest of the query code.
