+++ mypy.ini
@@ -82,7 +82,7 @@ warn_unused_ignores = False
 # ...except the modules and subpackages below (can't find a way to do line
 # breaks in the lists of modules).
 
-[mypy-lsst.daf.butler.registry.queries.expressions.parser.ply.*]
+[mypy-lsst.daf.butler.queries.expressions.parser.ply.*]
 ignore_errors = True
 
 [mypy-lsst.daf.butler.registry.tests.*]
