+++ pyproject.toml
@@ -204,7 +204,7 @@ extend-select = [
 
 [tool.ruff.lint.per-file-ignores]
 # parserYacc docstrings can not be fixed. Docstrings are used to define grammar.
-"python/lsst/daf/butler/registry/queries/expressions/parser/parserYacc.py" = ["D401", "D403"]
+"python/lsst/daf/butler/queries/expressions/parser/parserYacc.py" = ["D401", "D403"]
 
 [tool.ruff.lint.isort]
 known-first-party = ["lsst"]
