+++ python/lsst/daf/butler/queries/expressions/categorize.py
