+++ python/lsst/daf/butler/queries/expressions/parser/ply/README.md
