+++ python/lsst/daf/butler/queries/expressions/parser/ply/lex.py
