+++ tests/test_exprParserLex.py
@@ -30,7 +30,7 @@
 import re
 import unittest
 
-from lsst.daf.butler.registry.queries.expressions.parser import ParserLex, ParserLexError
+from lsst.daf.butler.queries.expressions.parser import ParserLex, ParserLexError
 
 
 class ParserLexTestCase(unittest.TestCase):
